+++ packages/DataElement/src/test/index.spec.ts
@@ -1,0 +1,14 @@
+import { describe, it, expect } from "@jest/globals";
+import {
+    DataElement,
+    StateChange,
+    customDataElements
+} from "../index";
+
+describe("DataElement index", () => {
+    it("exports the correct files", () => {
+        expect(DataElement).toBeDefined();
+        expect(StateChange).toBeDefined();
+        expect(customDataElements).toBeDefined();
+    });
+});