+++ packages/testUtils/src/index.ts
@@ -1,0 +1,5 @@
+export {
+    html,
+    fixture,
+    FixtureElement
+} from "./fixture";