+++ CHANGELOG.md
@@ -5,8 +5,61 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-### StateChange package to 0.6.0
+## [0.21.0] 2021-09-30
+### Updated
+- Router package 0.3.0
+
+## [0.20.0] 2021-09-29
+### Updated
+- Router package 0.2.0
+- DataElement 0.5.2
+- StateChange 0.7.2
+
+
+## [0.19.0] 2021-09-22
+### Added
+- Router package 0.1.1
+### Updated
+ - Event Map 0.7.1
+ - DataElement 0.5.1
+ - StateChange 0.7.1
+ - linkProp 0.3.1
+ - testUtils 0.2.1
+ - Middleware 0.2.9
+ - functional 0.2.7
+
+
+## [0.18.0] - 2021-09-10
+### Updated
+- badge inclusion and test coverage scripts
+- All packages patch version for documentation update
+
+## [0.17.0] - 2021-09-09
+### Updated
+- DataElements package to 0.4.0
+
+## [0.16.0] - 2021-09-08
+### Updated
+- DataElements package to 0.3.0
+
+## [0.15.0] - 2021-09-07
+### Updated
+- DataElements package to 0.2.0
+
+## [0.14.0] - 2021-09-06
+### Added
+- DataElements package 0.1.0
+
+## [0.13.0] - 2021-09-04
+### Added
+- testUtils package 0.1.0
+### Updated
+- EventMap package to 0.6.0
+
+
+## [0.12.0] - 2021-09-04
+### Updated
+- StateChange package to 0.6.0
 
 
 ## [0.11.0] - 2021-09-03